Improve stream usability.
When using binread with `:all` you could never get to the end of the
stream.  This way you can.. but I need more tests around dynamic files
to make sure we'll be able to read more once we hit `:eof`

Also, want to look into the selection of 4096 in `IO` and see if I can
tune my reads for better performance... eventually.